Optimize the layout for the notes
The notes two column layout is now optimized, so that both columns have the same height.
These columns are now independently scrollable. This layout now also works on mobile safari.
Also the button to create a new note has a totally new design and doesn't use the SubmitButton anymore.
And at last the icon and order for the notes page in the navigation bar and footer has been optimized.